+++ components/form/UploadCSV.tsx
@@ -23,6 +23,10 @@ import {
   hasExpectedColumnsERC721,
 } from "../../utils/csv";
 import TableCsvView from "./TableCsvView";
+import {
+  generateERC20SummaryData,
+  generateERC721SummaryData,
+} from "../../utils/generate_text";
 interface IReadData {
   fileParent?: File;
   setFile: (file: File) => void;
@@ -67,98 +71,6 @@ IReadData) => {
     }
   };
 
-  const generateSummaryData = (
-    tokensAddressSet: Set<string>,
-    datas: any[],
-    verifDataText: string
-  ) => {
-    let sumAmountByToken: Map<string, number> = new Map();
-    if (batchType == BatchType?.ERC20) {
-      for (let contractAddress of Array.from(tokensAddressSet.values())) {
-        sumAmountByToken[contractAddress] = 0;
-        sumAmountByToken.set(contractAddress, 0);
-        datas?.filter((callData) => {
-          if (callData?.token_address == contractAddress) {
-            let amount = Number(callData?.amount);
-            const value = sumAmountByToken.get(contractAddress);
-            sumAmountByToken.set(contractAddress, amount + value);
-            return callData?.amount;
-          }
-        });
-      }
-    }
-    // let summaryData = `Sum amount by tokens :\n`;
-    /** @TODO add React node */
-    let allText = Array.from(sumAmountByToken.keys()).map((key) => {
-      const value = sumAmountByToken.get(key);
-      const keyLen = key.length;
-      const truncAddress = `${key.slice(0, 10)}...${key.slice(
-        keyLen - 10,
-        keyLen
-      )} `;
-      let text = `${truncAddress} : ${value}\n`;
-      // summaryData +=text
-      return text;
-    });
-    const SummaryNode = (
-      <Box>
-        <Text>Sum amount by tokens </Text>
-        {allText?.map((t, i) => {
-          return <Text key={i}>{t}</Text>;
-        })}
-      </Box>
-    );
-
-    setSummaryNode(SummaryNode);
-    setSummaryData(summaryData);
-    setVerificationData(verifDataText);
-  };
-
-  const generateERC721SummaryData = (
-    tokensAddressSet: Set<string>,
-    datas: any[],
-    verifDataText: string
-  ) => {
-    let sumAmountByToken: Map<string, number> = new Map();
-    if (batchType == BatchType?.ERC721) {
-      for (let contractAddress of Array.from(tokensAddressSet.values())) {
-        sumAmountByToken[contractAddress] = 0;
-        sumAmountByToken.set(contractAddress, 0);
-        datas?.filter((callData) => {
-          if (callData?.token_address == contractAddress) {
-            const value = sumAmountByToken.get(contractAddress);
-            sumAmountByToken.set(contractAddress, 1 + value);
-            return callData?.amount;
-          }
-        });
-      }
-    }
-    // let summaryData = `Sum amount by tokens :\n`;
-    /** @TODO add React node */
-    let allText = Array.from(sumAmountByToken.keys()).map((key) => {
-      const value = sumAmountByToken.get(key);
-      const keyLen = key.length;
-      const truncAddress = `${key.slice(0, 10)}...${key.slice(
-        keyLen - 10,
-        keyLen
-      )} `;
-      let text = `${truncAddress} : ${value}\n`;
-      // summaryData +=text
-      return text;
-    });
-    const SummaryNode = (
-      <Box>
-        <Text>Sum amount by tokens </Text>
-        {allText?.map((t, i) => {
-          return <Text key={i}>{t}</Text>;
-        })}
-      </Box>
-    );
-
-    setSummaryNode(SummaryNode);
-    setSummaryData(summaryData);
-    setVerificationData(verifDataText);
-  };
   const readInfo = (datas?: any[]) => {
     try {
       console.log("readInfo", datas);
@@ -210,7 +122,15 @@ IReadData) => {
         setVerificationData(verifDataText);
 
         if (batchType == BatchType.ERC20) {
-          generateSummaryData(tokensAddressSet, datas, verifDataText);
+          const {verificationData, summaryData, summaryNode}= generateERC20SummaryData(
+            tokensAddressSet,
+            datas,
+            verifDataText,
+            batchType
+          );
+          setSummaryNode(summaryNode);
+          setSummaryData(summaryData);
+          setVerificationData(verificationData);
         }
 
         setCanTryBatch(true);
@@ -262,7 +182,16 @@ IReadData) => {
         setVerificationData(verifDataText);
 
         if (batchType == BatchType.ERC721) {
-          generateERC721SummaryData(tokensAddressSet, datas, verifDataText);
+          const { summaryNode, summaryData, verificationData } =
+            generateERC721SummaryData(
+              tokensAddressSet,
+              datas,
+              verifDataText,
+              batchType
+            );
+          setSummaryNode(summaryNode);
+          setSummaryData(summaryData);
+          setVerificationData(verificationData);
         }
 
         setCanTryBatch(true);
@@ -356,7 +285,6 @@ IReadData) => {
 
   return (
     <Box>
-    
       <Text fontSize={{ base: "15px", md: "19px" }}>
         {" "}
         Select if you want to batch ERC20 (token) or ERC721 (NFT)
