+++ components/modal/batch-tx/index.tsx
@@ -18,13 +18,11 @@ import TokenERC721Abi from "../../../constants/abi/erc721_token.json";
 import TokenERC20Abi from "../../../constants/abi/token_erc20.json";
 import {
   Call,
-  CallData,
   Contract,
   GetTransactionReceiptResponse,
   TransactionStatus,
   cairo,
   shortString,
-  uint256,
 } from "starknet";
 import { useState } from "react";
 import GearLoader from "../../loader/GearLoader";
@@ -32,6 +30,7 @@ import { ExternalStylizedButtonLink } from "../../button/NavItem";
 import { VoyagerExplorerImage } from "../../view/image/VoyagerExplorerImage";
 import { BatchType } from "../../../types";
 import { CHAINS_NAMES, CHAIN_IDS } from "../../../constants/address";
+import { prepareCallTransfer, prepareNftCallTransfer } from "../../../utils/callback";
 interface IBatchModal {
   modalOpen: boolean;
   chatId?: string;
@@ -49,13 +48,11 @@ interface IBatchModal {
 
 const BatchTxModal = ({
   modalOpen,
-  chatId,
   onClose,
   onOpen,
   restButton,
   csvData,
   verifData,
-  isDisabledModal,
   isCanTryBatch,
   batchType,
   summaryData,
@@ -81,84 +78,6 @@ const BatchTxModal = ({
   const calls: Call[] = [];
   const [callsData, setCallData] = useState<Call[]>([]);
 
-  /** Prepare call transfer
-   * Transform:Amount to uint256 with decimals
-   */
-  const prepareCallTransfer = async (
-    contractAddress: string,
-    recipient: string,
-    amount: number
-  ) => {
-    try {
-      // initialize contract using abi, address and provider
-      console.log("contractAddress", contractAddress);
-      console.log("recipient", recipient);
-      console.log("amount", amount);
-      const contract = new Contract(
-        TokenERC20Abi.abi,
-        contractAddress,
-        account
-      );
-      let decimals = 18;
-
-      try {
-        decimals = (await contract.decimals()) ?? 18;
-      } catch (e) {
-        console.log("decimals can be reach", e);
-      }
-      console.log("decimals", decimals);
-      let amountSent = Number(
-        Number(amount) * 10 ** Number(decimals)
-      ).toString();
-      const myCall: Call = contract.populate("transfer", [
-        recipient,
-        amountSent,
-      ]);
-      return myCall;
-    } catch (error: any) {
-      alert(error.message);
-    }
-  };
-
-  const prepareNftCallTransfer = async (
-    contractAddress: string,
-    recipient: string,
-    token_id: number
-  ) => {
-    try {
-      // initialize contract using abi, address and provider
-      console.log("contractAddress", contractAddress);
-      console.log("recipient", recipient);
-      console.log("token_id", token_id);
-
-      const contract = new Contract(
-        TokenERC721Abi.abi,
-        contractAddress,
-        account
-      );
-
-      // let call = {
-      //   contractAddress: contractAddress,
-      //   entrypoint: "transfer_from",
-      //   calldata: CallData.compile({
-      //     from: account?.address,
-      //     to: recipient, // to
-      //     token_id: token_id,
-      //   }),
-      // };
-
-      const call: Call = contract.populate("transfer_from", [
-        account?.address,
-        recipient,
-        token_id
-      ]);
-      return call;
-    } catch (error) {
-      console.log("error", error);
-      alert(error.message);
-      return undefined;
-    }
-  };
   // prepare each call data to transfer amount of token to the recipient
   const prepareTx = async () => {
     try {
@@ -211,7 +130,8 @@ const BatchTxModal = ({
           call = await prepareCallTransfer(
             tokenAddress,
             recipient,
-            Number(amount)
+            Number(amount),
+            account
           );
           decimals = (await contract.decimals()) ?? 18;
           console.log("call ERC20", call);
@@ -219,9 +139,9 @@ const BatchTxModal = ({
           call = await prepareNftCallTransfer(
             tokenAddress,
             recipient,
-            Number(token_id)
+            Number(token_id),
+            account
           );
-          console.log("call ERC721", call);
         }
         console.log("call", call);
